# internal/platform/database/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package sqlc

import (
	"net/netip"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type AccountActionToken struct {
	ID        uuid.UUID          `json:"id"`
	UserID    uuid.UUID          `json:"user_id"`
	Purpose   string             `json:"purpose"`
	TokenHash []byte             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type AuditEvent struct {
	ID         uuid.UUID          `json:"id"`
	ActorID    *uuid.UUID         `json:"actor_id"`
	Action     string             `json:"action"`
	Resource   string             `json:"resource"`
	ResourceID *uuid.UUID         `json:"resource_id"`
	Metadata   []byte             `json:"metadata"`
	IpAddress  *netip.Addr        `json:"ip_address"`
	UserAgent  string             `json:"user_agent"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type Availability struct {
	ID                    uuid.UUID          `json:"id"`
	ProfessionalProfileID uuid.UUID          `json:"professional_profile_id"`
	Weekday               int16              `json:"weekday"`
	StartTime             pgtype.Time        `json:"start_time"`
	EndTime               pgtype.Time        `json:"end_time"`
	CreatedAt             pgtype.Timestamptz `json:"created_at"`
}

type ClientProfile struct {
	ID          uuid.UUID          `json:"id"`
	UserID      uuid.UUID          `json:"user_id"`
	CompanyName string             `json:"company_name"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type PortfolioItem struct {
	ID                    uuid.UUID          `json:"id"`
	ProfessionalProfileID uuid.UUID          `json:"professional_profile_id"`
	StorageKey            string             `json:"storage_key"`
	OriginalName          string             `json:"original_name"`
	MediaType             string             `json:"media_type"`
	SizeBytes             int64              `json:"size_bytes"`
	SortOrder             int16              `json:"sort_order"`
	CreatedAt             pgtype.Timestamptz `json:"created_at"`
}

type ProfessionalProfile struct {
	ID                   uuid.UUID          `json:"id"`
	UserID               uuid.UUID          `json:"user_id"`
	CreatedAt            pgtype.Timestamptz `json:"created_at"`
	UpdatedAt            pgtype.Timestamptz `json:"updated_at"`
	Bio                  string             `json:"bio"`
	PrimaryCategoryID    *uuid.UUID         `json:"primary_category_id"`
	Skills               []string           `json:"skills"`
	ServiceRegion        string             `json:"service_region"`
	ServiceLocation      interface{}        `json:"service_location"`
	ServiceRadiusKm      int32              `json:"service_radius_km"`
	ReferencePriceCents  *int64             `json:"reference_price_cents"`
	AvailableNow         bool               `json:"available_now"`
	AvailabilityTimezone string             `json:"availability_timezone"`
}

type RefreshToken struct {
	ID        uuid.UUID          `json:"id"`
	UserID    uuid.UUID          `json:"user_id"`
	TokenHash []byte             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	RevokedAt pgtype.Timestamptz `json:"revoked_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type ServiceCategory struct {
	ID          uuid.UUID          `json:"id"`
	Slug        string             `json:"slug"`
	Name        string             `json:"name"`
	Description string             `json:"description"`
	IsActive    bool               `json:"is_active"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type SystemLocation struct {
	ID        uuid.UUID          `json:"id"`
	Label     string             `json:"label"`
	Point     interface{}        `json:"point"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID              uuid.UUID          `json:"id"`
	Email           string             `json:"email"`
	PasswordHash    string             `json:"password_hash"`
	DisplayName     string             `json:"display_name"`
	EmailVerifiedAt pgtype.Timestamptz `json:"email_verified_at"`
	Status          string             `json:"status"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
}