# openapi.yaml
openapi: 3.0.3
info:
  title: Tucano API
  version: 0.3.0
  description: >-
    API do marketplace de profissionais autônomos. A plataforma facilita descoberta,
    compatibilidade e contato direto; não processa o pagamento dos serviços.
servers:
  - url: /api
paths:
...
  /auth/register:
    post:
      tags: [authentication]
      summary: Cria conta com o primeiro perfil
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/RegisterRequest"}}}
      responses:
        "201": {description: Conta criada; confirmação de e-mail pendente}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /auth/verify-email:
    post:
      tags: [authentication]
      summary: Confirma o e-mail com token de uso único
      requestBody: {$ref: "#/components/requestBodies/Token"}
      responses:
        "200": {description: E-mail confirmado}
        "401": {$ref: "#/components/responses/InvalidToken"}
  /auth/resend-verification:
    post:
      tags: [authentication]
      summary: Solicita novo link sem revelar se a conta existe
      requestBody: {$ref: "#/components/requestBodies/Email"}
      responses:
        "202": {description: Solicitação aceita}
  /auth/login:
    post:
      tags: [authentication]
      summary: Autentica e define o refresh token em cookie HttpOnly
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/LoginRequest"}}}
      responses:
        "200":
          description: Sessão criada
          headers:
            Set-Cookie: {schema: {type: string}, description: Refresh token HttpOnly e SameSite=Strict}
          content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}
        "401": {description: Credenciais inválidas, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "403": {description: E-mail pendente ou conta indisponível, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /auth/refresh:
    post:
      tags: [authentication]
      summary: Rotaciona o refresh token e emite novo access token
      responses:
        "200": {description: Sessão renovada, content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}}
        "401": {$ref: "#/components/responses/InvalidToken"}
...
  /auth/forgot-password:
    post:
      tags: [authentication]
      summary: Solicita redefinição sem revelar se a conta existe
      requestBody: {$ref: "#/components/requestBodies/Email"}
      responses:
        "202": {description: Solicitação aceita}
  /auth/reset-password:
    post:
      tags: [authentication]
      summary: Redefine a senha e revoga todas as sessões
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ResetPasswordRequest"}}}
      responses:
        "200": {description: Senha redefinida}
        "401": {$ref: "#/components/responses/InvalidToken"}
  /account/me:
    get:
      tags: [account]
      summary: Retorna a conta e seus perfis atuais
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Conta atual}
        "401": {$ref: "#/components/responses/InvalidToken"}
  /account/profiles:
    post:
      tags: [account]
      summary: Adiciona o segundo perfil à mesma conta
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ProfileRequest"}}}
      responses:
        "201": {description: Perfil adicionado}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
    Email:
      required: true
      content: {application/json: {schema: {type: object, required: [email], properties: {email: {type: string, format: email}}}}}
    Token:
      required: true
      content: {application/json: {schema: {type: object, required: [token], properties: {token: {type: string}}}}}
  responses:
    InvalidToken:
      description: Token ausente, inválido ou expirado
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    ValidationError:
      description: Dados inválidos
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Conflict:
      description: Recurso já existente
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, display_name, profile_type]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, format: password, minLength: 12, maxLength: 128}
        display_name: {type: string, minLength: 2, maxLength: 100}
        profile_type: {type: string, enum: [professional, client]}
        company_name: {type: string, maxLength: 150}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, format: password}
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token: {type: string}
        password: {type: string, format: password, minLength: 12, maxLength: 128}
    ProfileRequest:
      type: object
      required: [profile_type]
      properties:
        profile_type: {type: string, enum: [professional, client]}
        company_name: {type: string, maxLength: 150}
...
    User:
      type: object
      required: [id, email, display_name, email_verified, roles]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        display_name: {type: string}
        email_verified: {type: boolean}
        roles:
          type: array
          items: {type: string, enum: [professional, client]}
    Session:
      type: object
      required: [access_token, access_expires_at, user]
      properties:
        access_token: {type: string}
        access_expires_at: {type: string, format: date-time}
        user: {$ref: "#/components/schemas/User"}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}